#13 'if' statement with common parts inspection   2 weak warnings #13 Redundant 'if' statement inspection   4 warnings #13 Simplifiable conditional expression inspection   1 warning #13 Conditional can be pushed inside branch expression inspection   33 warnings #13 Conditional expression with negated condition inspection   1 warning #13 Negated equality expression inspection   6 warnings

+++ src/main/java/gov/nasa/gsfc/spdf/cdfj/FloatVarContainer.java
@@ -170,7 +170,7 @@ public final class FloatVarContainer extends BaseVarContainer implements
                 for (int e = 0; e < _num; e++) {
                     Number num = (Number)method.invoke(bv, new Object[] {});
                     int x = num.intValue();
-                    data[e] = (x >= 0)?(float)x:(float)(longInt + x);
+                    data[e] = (float) (x >= 0 ? x : longInt + x);
                 }
                 fbuf.put(data, 0, _num);
                 position += 4*_num;
